Comment out empty values in config files.
Entries of the form "foo=" only work if IMP is built with a
recent version of Boost (1.56 or later) due to a bug in older
versions of Boost (see boostorg/program_options@ccbbcab336).
One workaround which works in all Boost versions is just to comment
out (or remove) such entries, so they get the default value (which
is blank).